health-connect-ai/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
aioredis==2.0.1
rasa==3.6.9
scispacy==0.5.5
pytest==8.3.5